runpod/requirements.txt
# RunPod Requirements - 修复模块缺失问题
runpod==1.5.1
llama-cpp-python>=0.2.11
numpy>=1.24.3
torch>=2.0.1
requests>=2.31.0

# 兼容性依赖
setuptools>=65.0.0
wheel>=0.37.0 